# openapi.yml
openapi: 3.0.3
info:
  title: Order Service API
  description: |
    Microservice for managing customer orders.
    Communicates with Product Service to validate products and reserve stock.
  version: 1.0.0

servers:
  - url: http://localhost:8082
    description: Local development
  - url: http://<your-ecs-public-ip>:8082
    description: AWS ECS deployment

tags:
  - name: Orders
    description: Order management endpoints

paths:
  /api/orders:
    get:
      tags: [Orders]
      summary: Get all orders
      responses:
        '200':
          description: List of all orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      tags: [Orders]
      summary: Create a new order
      description: |
        Creates an order by:
        1. Calling Product Service GET /api/products/{id} to validate each product
        2. Checking stock availability
        3. Calling Product Service POST /api/products/{id}/reserve to reserve stock
        4. Persisting the confirmed order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
            example:
              customerName: "John Doe"
              customerEmail: "john@example.com"
              items:
                - productId: 1
                  quantity: 2
                - productId: 3
                  quantity: 1
      responses:
        '201':
          description: Order created and confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error or product not found
        '409':
          description: Insufficient stock
        '503':
          description: Product Service unavailable

  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

  /api/orders/customer/{email}:
    get:
      tags: [Orders]
      summary: Get orders by customer email
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Orders for customer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

  /api/orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [PENDING, CONFIRMED, CANCELLED, DELIVERED]
      responses:
        '200':
          description: Status updated
        '404':
          description: Order not found

  /api/orders/health:
    get:
      tags: [Orders]
      summary: Health check
      responses:
        '200':
          description: Service is healthy

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, DELIVERED]
        totalAmount:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        createdAt:
          type: string
          format: date-time

    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number

    CreateOrderRequest:
      type: object
      required: [customerName, customerEmail, items]
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
          format: email
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
                minimum: 1